requirements.txt
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
google-cloud-logging